Changes according to PR review
-Created a more structured version
-Login and SignUp feature are now in different files
