+++ src/components/PublicHomePage/PublicHomePage.css
@@ -25,3 +25,14 @@
     background-color: #ff964f;
   }
   
+  .login-btn {
+    padding: 12px 30px;
+    font-size: 1.2rem;
+    font-weight: bold;
+    border-radius: 25px;
+    transition: background 0.3s ease-in-out;
+  }
+  
+  .login-btn:hover {
+    background-color: #ff964f;
+  }
