+++ src/components/Signup/SignUp.css
@@ -1,0 +1,47 @@
+.auth-container {
+    min-height: 100vh;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background: url('/public/auth\ bg.jpg') no-repeat center center/cover;
+    position: relative;
+  }
+  
+  .button-group {
+    display: flex;
+    gap: 1.5rem;
+    justify-content: center; 
+    margin-top: 2vh; 
+  }
+
+  .auth-container::before {
+    content: "";
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background: rgba(0, 0, 0, 0.5); 
+  }
+  
+  .auth-card {
+    position: relative;
+    width: 30rem;
+    background: rgba(255, 255, 255, 0.9); 
+    box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.2);
+    padding: 20px;
+    border-radius: 10px;
+  }
+
+  .auth-button {
+    background-color: #4e6ba6 !important;
+    border-color: #4e6ba6 !important;
+    color: #fff !important; 
+  }
+  
+  .auth-button:hover {
+    background-color: #94B6C1 !important; 
+    border-color: #94B6C1 !important;
+  }
+  
+
